# swagger-collection.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apartment Management API v1.0",
    "version": "1.0.0",
    "description": "OpenAPI 3 — import vào Swagger Editor / UI. Mọi JSON body dùng schema có `example` trên từng field. Các UUID trong ví dụ có dạng ngẫu nhiên (giống `Guid.NewGuid()` khi seed) và **không** phải id thật trong DB — sau khi seed, thay bằng giá trị từ API. JWT: login rồi gửi Authorization: Bearer. Seed data range: apartments BLK-0001..BLK-0150, bulk users demoblk00..demoblk39@seed.local, canonical accounts admin1/admin2 and user1..user6 @ apartment.local. UUID examples are placeholders; after seeding, lấy ID thật từ API list/detail."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1.0/Invoices/me": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "My invoices",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "example": 10
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": ""
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "createdAt"
          },
          {
            "name": "descending",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "example": false
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Invoices/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Get invoice (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Invoices"
        ],
        "summary": "Update invoice (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Invoices"
        ],
        "summary": "Delete invoice (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Invoices/{id}/restore": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "post": {
        "tags": [
          "Invoices"
        ],
        "summary": "Restore invoice (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1.0/Feedbacks/me": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "My feedback posts",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "example": 10
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": ""
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "createdAt"
          },
          {
            "name": "descending",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "example": false
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Feedbacks/tree": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "Feedback tree",
        "parameters": [
          {
            "name": "rootFeedbackId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "example": null
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Feedbacks/flattened": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "Feedback flattened",
        "parameters": [
          {
            "name": "rootFeedbackId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "example": null
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Feedbacks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "Get feedback by id (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "Delete feedback",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Feedbacks/{id}/restore": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "post": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "Restore feedback (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "InvoiceDetailCreateDto": {
        "type": "object",
        "required": [
          "utilityServiceId",
          "quantity",
          "unitPrice"
        ],
        "properties": {
          "utilityServiceId": {
            "type": "string",
            "format": "uuid",
            "example": "6b1e3d8c-0f2a-4b9e-a156-3e4f5a6b7890"
          },
          "quantity": {
            "type": "number",
            "example": 24
          },
          "unitPrice": {
            "type": "number",
            "example": 2200
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Line 12"
          }
        }
      },
      "InvoiceCreateDto": {
        "type": "object",
        "required": [
          "apartmentId",
          "title",
          "billingMonth",
          "issueDate",
          "dueDate",
          "isPaid",
          "details"
        ],
        "properties": {
          "apartmentId": {
            "type": "string",
            "format": "uuid",
            "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
          },
          "title": {
            "type": "string",
            "example": "Invoice 2025-03 — unit 150"
          },
          "billingMonth": {
            "type": "string",
            "format": "date",
            "example": "2025-03-01"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-05T00:00:00.000Z"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-19T00:00:00.000Z"
          },
          "isPaid": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDetailCreateDto"
            },
            "example": [
              {
                "utilityServiceId": "6b1e3d8c-0f2a-4b9e-a156-3e4f5a6b7890",
                "quantity": 24,
                "unitPrice": 2200,
                "note": "Line 12"
              }
            ]
          }
        }
      },
      "InvoiceUpdateDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/InvoiceCreateDto"
          }
        ]
      },
...
    }
  }
}